% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loveletters.R
\name{loveletters}
\alias{loveletters}
\title{Love letter creator}
\usage{
loveletters(recipient, sender, descriptor)
}
\arguments{
\item{recipient}{who do you love (str)?}

\item{sender}{who do you think you are (str)?}

\item{descriptor}{a word or phrase describing the recipient (srt)!}
}
\description{
Love letter creator
}
\examples{
loveletters("Jillian", "Jillian", "a coding genius")

}
